change the type of id to Long

+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvAbilityGenerations.java
@@ -3,6 +3,6 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvAbilityGenerations(
-    @CsvBindByName(column = "ability_id") int abilityId,
-    @CsvBindByName(column = "generation_id") int generationId
+    @CsvBindByName(column = "ability_id") Long abilityId,
+    @CsvBindByName(column = "generation_id") Long generationId
 ) {}
