+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvExperience.java
@@ -3,7 +3,7 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvExperience(
-    @CsvBindByName(column = "growth_rate_id") int growthRateId,
+    @CsvBindByName(column = "growth_rate_id") Long growthRateId,
     @CsvBindByName(column = "level") int level,
     @CsvBindByName(column = "experience") long experience
 ) {}
