+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvItemCategories.java
@@ -3,8 +3,8 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvItemCategories(
-    @CsvBindByName(column = "id") int id,
-    @CsvBindByName(column = "pocket_id") int pocketId,
+    @CsvBindByName(column = "id") Long id,
+    @CsvBindByName(column = "pocket_id") Long pocketId,
     @CsvBindByName(column = "identifier") String identifier,
     @CsvBindByName(column = "name") String name
 ) {}
