+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvItemPockets.java
@@ -3,7 +3,7 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvItemPockets(
-    @CsvBindByName(column = "id") int id,
+    @CsvBindByName(column = "id") Long id,
     @CsvBindByName(column = "identifier") String identifier,
     @CsvBindByName(column = "name") String name
 ) {}
