+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvMove.java
@@ -3,16 +3,16 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvMove(
-		@CsvBindByName(column = "id") int id,
+		@CsvBindByName(column = "id") Long id,
 		@CsvBindByName(column = "identifier") String identifier,
-		@CsvBindByName(column = "generation_id") int generationId,
-		@CsvBindByName(column = "type_id") int typeId,
+		@CsvBindByName(column = "generation_id") Long generationId,
+		@CsvBindByName(column = "type_id") Long typeId,
 		@CsvBindByName(column = "power") Double power,
 		@CsvBindByName(column = "pp") Integer pp,
 		@CsvBindByName(column = "accuracy") Integer accuracy,
 		@CsvBindByName(column = "priority") int priority,
-		@CsvBindByName(column = "target_id") int targetId,
-		@CsvBindByName(column = "damage_class_id") int damageClassId,
-		@CsvBindByName(column = "effect_id") int effectId,
+		@CsvBindByName(column = "target_id") Long targetId,
+		@CsvBindByName(column = "damage_class_id") Long damageClassId,
+		@CsvBindByName(column = "effect_id") Long effectId,
 		@CsvBindByName(column = "name") String name
 ) {}
