+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvMoveChangelog.java
@@ -3,9 +3,9 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvMoveChangelog(
-		@CsvBindByName(column = "move_id") int moveId,
-		@CsvBindByName(column = "changed_in_version_group_id") int versionGroupId,
-		@CsvBindByName(column = "type_id") int typeId,
+		@CsvBindByName(column = "move_id") Long moveId,
+		@CsvBindByName(column = "changed_in_version_group_id") Long versionGroupId,
+		@CsvBindByName(column = "type_id") Long typeId,
 		@CsvBindByName(column = "power") Integer power,
 		@CsvBindByName(column = "pp") Integer pp,
 		@CsvBindByName(column = "accuracy") Integer accuracy,
