+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvPokedexVersionGroups.java
@@ -3,6 +3,6 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvPokedexVersionGroups(
-    @CsvBindByName(column = "pokedex_id") int pokedexId,
-    @CsvBindByName(column = "version_group_id") int versionGroupId
+    @CsvBindByName(column = "pokedex_id") Long pokedexId,
+    @CsvBindByName(column = "version_group_id") Long versionGroupId
 ) {}
