+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvPokedexes.java
@@ -3,8 +3,8 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvPokedexes(
-    @CsvBindByName(column = "id") int id,
-    @CsvBindByName(column = "region_id") int regionId,
+    @CsvBindByName(column = "id") Long id,
+    @CsvBindByName(column = "region_id") Long regionId,
     @CsvBindByName(column = "identifier") String identifier,
     @CsvBindByName(column = "is_main_series") int isMainSeries
 ) {}
