+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvPokemon.java
@@ -3,9 +3,9 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvPokemon(
-		@CsvBindByName(column = "id") int id,
+		@CsvBindByName(column = "id") Long id,
 		@CsvBindByName(column = "identifier") String identifier,
-		@CsvBindByName(column = "species_id") int speciesId,
+		@CsvBindByName(column = "species_id") Long speciesId,
 		@CsvBindByName(column = "height") int height,
 		@CsvBindByName(column = "weight") int weight,
 		@CsvBindByName(column = "base_experience") int baseExperience,
