+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvSuperContestCombos.java
@@ -3,6 +3,6 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvSuperContestCombos(
-    @CsvBindByName(column = "first_move_id") int firstMoveId,
-    @CsvBindByName(column = "second_move_id") int secondMoveId
+    @CsvBindByName(column = "first_move_id") Long firstMoveId,
+    @CsvBindByName(column = "second_move_id") Long secondMoveId
 ) {}
