+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvTypes.java
@@ -3,8 +3,8 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvTypes(
-    @CsvBindByName(column = "id") int id,
+    @CsvBindByName(column = "id") Long id,
     @CsvBindByName(column = "identifier") String identifier,
-    @CsvBindByName(column = "generation_id") int generationId,
-    @CsvBindByName(column = "damage_class_id") int damageClassId
+    @CsvBindByName(column = "generation_id") Long generationId,
+    @CsvBindByName(column = "damage_class_id") Long damageClassId
 ) {}
