+++ src/main/java/at/jku/faw/neo4jdemo/model/csv/CsvVersionGroup.java
@@ -3,9 +3,9 @@ package at.jku.faw.neo4jdemo.model.csv;
 import com.opencsv.bean.CsvBindByName;
 
 public record CsvVersionGroup(
-		@CsvBindByName(column = "id") int id,
+		@CsvBindByName(column = "id") Long id,
 		@CsvBindByName(column = "identifier") String identifier,
-		@CsvBindByName(column = "generation_id") int generationId,
+		@CsvBindByName(column = "generation_id") Long generationId,
 		@CsvBindByName(column = "order") int order
 ) {
 }
